refactor: Move utils to `_main_utils.py`
